Add Radix Accordion component and update UI elements
